Added hints on how to finish day 2

+++ AdventOfCode/Day6/ILanternfishes.cs
@@ -2,7 +2,7 @@
 {
     public interface ILanternfishes
     {
-        int Count { get; }
+        long Count { get; }
 
         void SimulateOneDay();
     }
